Simplify the process of creating new test plans
Remove the boilerplate required to create new test plans. This of coarse
has the side-effect of using Ruby-magic, but it's only a little so it
seemed worth it.

The caveat is that the inference of the class/file name is very naive
and it could be better. I didn't really feel like pulling ActiveSupport
or re-writing the constantize/camelize methods myself, so I've left it
alone.